ocr_module/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
pytesseract
pdf2image
pyzbar